+++ libmfpic/ElectrostaticFieldOperationsWithBoltzmannElectrons.hpp
@@ -13,7 +13,7 @@ public:
     double electron_reference_number_density,
     double electron_temperature,
     double nonlinear_solver_relative_tolerance = 1.0e-8,
-    double nonlinear_solver_max_iterations = 100
+    int nonlinear_solver_max_iterations = 100
   );
 
   virtual void fieldSolve(ElectrostaticFieldState& field_state, const IntegratedCharge& charge_state) override;
